requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
psycopg2==2.9.11  # Bu yangi versiya o'rnatildi
SQLAlchemy==2.0.19
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2